+++ src/main/java/kr/co/webee/application/hive/handler/TelemetryHandler.java
@@ -1,14 +1,16 @@
-package kr.co.webee.application.mqtt;
+package kr.co.webee.application.hive.handler;
 
-import kr.co.webee.application.hive.dto.HiveTelemetryRequest;
+import kr.co.webee.application.hive.dto.request.HiveTelemetryRequest;
 import kr.co.webee.application.hive.service.HiveTelemetryService;
+import kr.co.webee.application.mqtt.MqttMessageHandler;
+import kr.co.webee.application.mqtt.MqttTopicType;
 import kr.co.webee.common.util.JsonConverter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
 @Component
-public class TelemetryHandler implements MqttMessageHandler{
+public class TelemetryHandler implements MqttMessageHandler {
     private final HiveTelemetryService hiveTelemetryService;
     private final JsonConverter jsonConverter;
